feature: resolve merge conflicts
